game exit button works
