+++ web/src/pages/Game/GameCanvas/styles.scss
@@ -1,0 +1,1 @@
+
