+++ web/src/pages/Main/FindGame/styles.scss
@@ -11,9 +11,10 @@
 	z-index: 9;
 
 	.find-game-img {
+		margin-top: 5px;
 		width: 100%;
-		height: 130px;
-		background-image: url("../../../img/pong-logo.png");
+		height: 75px;
+		background-image: url("../../../img/play-pong.png");
 		background-size: contain;
 		background-repeat: no-repeat;
 		background-position: center;
@@ -24,6 +25,7 @@
 	}
 
 	.find-game-back {
+		margin-bottom: 5px;
 		width: 370px;
 		display: flex;
 		justify-content: space-between;
